fix:code refactoring of transit_gateway file
