+++ aws/resources/tgw_vpc_attachment_types.go
@@ -11,38 +11,34 @@ import (
 	"github.com/gruntwork-io/go-commons/errors"
 )
 
-// TransitGateways - represents all transit gateways
-type TransitGateways struct {
+// TransitGatewaysVpcAttachment - represents all transit gateways vpc attachments
+type TransitGatewaysVpcAttachment struct {
 	BaseAwsResource
 	Client ec2iface.EC2API
 	Region string
 	Ids    []string
 }
 
-func (tgw *TransitGateways) Init(session *session.Session) {
-	// to initialize base resource
-	// NOTE : This is madatory to initialize the nukables map
-	tgw.BaseAwsResource.Init(session)
+func (tgw *TransitGatewaysVpcAttachment) Init(session *session.Session) {
 	tgw.Client = ec2.New(session)
-
 }
 
 // ResourceName - the simple name of the aws resource
-func (tgw *TransitGateways) ResourceName() string {
-	return "transit-gateway"
+func (tgw *TransitGatewaysVpcAttachment) ResourceName() string {
+	return "transit-gateway-attachment"
 }
 
 // MaxBatchSize - Tentative batch size to ensure AWS doesn't throttle
-func (tgw *TransitGateways) MaxBatchSize() int {
+func (tgw *TransitGatewaysVpcAttachment) MaxBatchSize() int {
 	return maxBatchSize
 }
 
 // ResourceIdentifiers - The Ids of the transit gateways
-func (tgw *TransitGateways) ResourceIdentifiers() []string {
+func (tgw *TransitGatewaysVpcAttachment) ResourceIdentifiers() []string {
 	return tgw.Ids
 }
 
-func (tgw *TransitGateways) GetAndSetIdentifiers(c context.Context, configObj config.Config) ([]string, error) {
+func (tgw *TransitGatewaysVpcAttachment) GetAndSetIdentifiers(c context.Context, configObj config.Config) ([]string, error) {
 	identifiers, err := tgw.getAll(c, configObj)
 	if err != nil {
 		return nil, err
@@ -53,7 +49,7 @@ func (tgw *TransitGateways) GetAndSetIdentifiers(c context.Context, configObj co
 }
 
 // Nuke - nuke 'em all!!!
-func (tgw *TransitGateways) Nuke(identifiers []string) error {
+func (tgw *TransitGatewaysVpcAttachment) Nuke(identifiers []string) error {
 	if err := tgw.nukeAll(awsgo.StringSlice(identifiers)); err != nil {
 		return errors.WithStackTrace(err)
 	}
